% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diatbarcode.R
\name{download_diatbarcode}
\alias{download_diatbarcode}
\title{Download the diat.barcode database}
\usage{
download_diatbarcode(path = NULL, flavor = "original",
  version = "last")
}
\arguments{
\item{path}{the path where the downloaded file is to be saved. If \code{NULL}, the temporary directory.}

\item{flavor}{a flavor of the Diat.barcode database (see Details section).}

\item{version}{a character string giving the version. Use \code{"last"} (default) to get the last version available.}
}
\value{
Silently returns a list with informations about the downloaded file.
}
\description{
This function downloads the database in a given location on your computer.
The function allows to download different flavors of the original database.
}
\details{
The Diat.barcode database can be downloaded in different flavors. Currently available flavors are:
\itemize{
  \item \code{original} The original database.
  \item \code{rbcl312_dada2_tax} A pre-formated reference for taxonomic affiliation with DADA2.
  \item \code{rbcl312_dada2_spe} A pre-formated reference for exact species matching with DADA2.
}

The diat.barcode database is a library of barcodes which is curated and
which guarantees good taxonomical homogeneity and good quality of sequences.
Data have different sources:
\itemize{
  \item The culture collection of Thonon, INRA, TCC.
  \item The barcoding project of diatoms in UK.
  \item Public data (NCBI).
}
Several European labs participate in this curation:
\itemize{
  \item INRA – UMR Carrtel, FR–74200 Thonon–les–Bains, France.
  \item Botanischer Garten und Botanisches Museum Berlin–Dahlem, Freie Universität Berlin, Germany.
  \item Swedish University of Agricultural Sciences, Department of Aquatic Sciences and Assessment, Uppsala, Sweden.
  \item Bowburn Consultancy, Durham, UK.
  \item Royal Botanic Garden Edinburgh, Edinburgh, Scotland, UK.
  \item Marine and Continental Waters, Institute for Food and Agricultural Research and Technology (IRTA), Catalonia, Spain.
  \item Laboratory for Evolutionary Ecology, center for Marine Research, Ruder Boskovic Institute, Rovinj, Croatia.
}
}
\section{Licence}{

Diatbarcode and R-Syst are distributed under the terms of the Open Licence 2.0.
Consult: \url{https://www.etalab.gouv.fr/wp-content/uploads/2018/11/open-licence.pdf}
}

\references{
Rimet, Frederic; Gusev, Evgenuy; Kahlert, Maria; Kelly, Martyn; Kulikovskiy, Maxim; Maltsev,
Yevhen; Mann, David; Pfannkuchen, Martin; Trobajo, Rosa; Vasselon, Valentin; Zimmermann, Jonas;
Bouchez, Agnès, 2018, "Diat.barcode, an open-access barcode library for diatoms",
\url{https://doi.org/10.15454/TOMBYZ}, Portail Data Inra, V1

Rimet F., Chaumeil P., Keck F., Kermarrec L., Vasselon V., Kahlert M., et al. (2016).
R-Syst::diatom: an open-access and curated barcode database for diatoms and freshwater monitoring.
Database 2016, baw016. \url{https://doi.org/10.1093/database/baw016}
}
