% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diatbarcode.R
\name{get_diatbarcode}
\alias{get_diatbarcode}
\title{Get the diat.barcode database
This function get the \code{diat.barcode} database from the official server.}
\usage{
get_diatbarcode(version = "last", verbose = TRUE)
}
\arguments{
\item{version}{a character string giving the version. Use \code{"last"} (default) to get the last version available.}

\item{verbose}{a logical. Set to \code{FALSE} hide the automatic message. Default is \code{TRUE}.}
}
\value{
A tibble object.
}
\description{
Get the diat.barcode database
This function get the \code{diat.barcode} database from the official server.
}
